user linked, ckeditor, cover_image_upload

+++ app/Http/Controllers/PostController.php
@@ -8,6 +8,12 @@ use App\Models\Category;
 
 class PostController extends Controller
 {
+
+     public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -57,8 +63,25 @@ class PostController extends Controller
         $this->validate($request, [
             'title' => 'required|string',
             'body' => 'required|string',
-            'cat_id' => 'required|integer'
+            'cat_id' => 'required|integer',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+         if($request->hasFile('cover_image')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+		
+	   } else {
+            $fileNameToStore = 'default.jpg';
+        }
        
 
         // Create a new post
@@ -66,9 +89,11 @@ class PostController extends Controller
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->cat_id =$request->input('cat_id');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
-        return redirect()->route('posts.index');
+        return redirect()->route('posts.index')->with('success', 'Post Created');
     }
 
     /**
@@ -95,6 +120,11 @@ class PostController extends Controller
      */
     public function edit($id)
     {
+
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         $categories = Category::all();
        $post = Post::find($id);
         return view('posts.edit')->with('post', $post);
@@ -117,12 +147,29 @@ class PostController extends Controller
             'cat_id' => 'required|integer'
         ]);
 
+        if($request->hasFile('cover_image')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+            // Delete file if exists
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post = Post::find($id);
           $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->cat_id =$request->input('cat_id');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
-         return redirect()->route('posts.index');
+         return redirect()->route('posts.index')->with('success', 'Post Updated');
     }
 
     /**
@@ -133,9 +180,16 @@ class PostController extends Controller
      */
     public function destroy($id)
     {
-    
+      if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+        if($post->cover_image != 'default.jpg'){
+            // Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
          $post = Post::find($id);
            $post->delete();
-        return redirect()->route('posts.index');
+        return redirect()->route('posts.index')->with('success', 'Post Removed');;
     }
 }
